# specs/002-fullstack-todo-app/contracts/api.yaml
openapi: 3.0.3
info:
  title: Full-Stack Todo Web Application API
  description: |
    REST API for secure, multi-user todo task management.

    **Security**: All endpoints (except auth) require JWT authentication via Bearer token.
    User isolation is enforced at the database query level.

    **Base URL**: `http://localhost:8000` (development)
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: Authentication
    description: User registration and login operations
  - name: Tasks
    description: Task CRUD operations (authenticated users only)

security:
  - BearerAuth: []

paths:
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Register new user account
      description: |
        Create a new user account with email and password.
        Returns JWT token upon successful registration.
      operationId: signup
      security: []  # No auth required for signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              email: user@example.com
              password: SecurePassword123
      responses:
        '201':
          description: User account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                email: user@example.com
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          description: Validation error (invalid email format, weak password, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Email already registered"
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login with existing account
      description: |
        Authenticate user with email and password.
        Returns JWT token upon successful login.
      operationId: login
      security: []  # No auth required for login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                email: user@example.com
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Invalid email or password"
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/{user_id}/tasks:
    get:
      tags:
        - Tasks
      summary: Get all tasks for authenticated user
      description: |
        Retrieve all tasks belonging to the authenticated user.
        Optionally filter by completion status.

        **Security**: JWT user_id must match path user_id (403 if mismatch).
      operationId: getTasks
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - name: completed
          in: query
          description: Filter tasks by completion status
          required: false
          schema:
            type: boolean
            example: false
      responses:
        '200':
          description: List of tasks retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
              example:
                - id: 1
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  title: "Complete project documentation"
                  description: "Write comprehensive API docs"
                  completed: false
                  created_at: "2026-02-08T10:30:00Z"
                  updated_at: "2026-02-08T10:30:00Z"
                - id: 2
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  title: "Review pull requests"
                  description: null
                  completed: true
                  created_at: "2026-02-07T14:20:00Z"
                  updated_at: "2026-02-08T09:15:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Tasks
      summary: Create new task
      description: |
        Create a new task for the authenticated user.
        Task is automatically associated with the user from JWT token.

        **Security**: JWT user_id must match path user_id (403 if mismatch).
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
            example:
              title: "Implement user authentication"
              description: "Add JWT-based auth to the API"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                id: 3
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Implement user authentication"
                description: "Add JWT-based auth to the API"
                completed: false
                created_at: "2026-02-08T11:00:00Z"
                updated_at: "2026-02-08T11:00:00Z"
        '400':
          description: Validation error (empty title, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Validation error"
                field_errors:
                  title: "Title cannot be empty"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/{user_id}/tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Get specific task by ID
      description: |
        Retrieve a single task by ID.

        **Security**: JWT user_id must match path user_id AND task owner (403 if mismatch).
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                id: 1
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Complete project documentation"
                description: "Write comprehensive API docs"
                completed: false
                created_at: "2026-02-08T10:30:00Z"
                updated_at: "2026-02-08T10:30:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Tasks
      summary: Update existing task
      description: |
        Update task fields (title, description, completed status).
        Only the task owner can update the task.

        **Security**: JWT user_id must match path user_id AND task owner (403 if mismatch).
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
            example:
              title: "Complete project documentation (updated)"
              description: "Write comprehensive API docs with examples"
              completed: true
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                id: 1
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Complete project documentation (updated)"
                description: "Write comprehensive API docs with examples"
                completed: true
                created_at: "2026-02-08T10:30:00Z"
                updated_at: "2026-02-08T12:45:00Z"
        '400':
          description: Validation error (empty title, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Validation error"
                field_errors:
                  title: "Title cannot be empty"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: |
        Permanently delete a task.
        Only the task owner can delete the task.

        **Security**: JWT user_id must match path user_id AND task owner (403 if mismatch).
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '204':
          description: Task deleted successfully (no content returned)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Task not found"
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /api/auth/signup or /api/auth/login.
        Include in Authorization header: `Authorization: Bearer <token>`

  parameters:
    UserIdPath:
      name: user_id
      in: path
      description: User ID (must match authenticated user's ID from JWT token)
      required: true
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"

    TaskIdPath:
      name: task_id
      in: path
      description: Task ID
      required: true
      schema:
        type: integer
        minimum: 1
        example: 1

  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 255
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 255
          description: User password (min 8 characters)
          example: SecurePassword123

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User password
          example: SecurePassword123

    AuthResponse:
      type: object
      required:
        - user_id
        - email
        - token
      properties:
        user_id:
          type: string
          format: uuid
          description: Unique user identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        token:
          type: string
          description: JWT token for authentication
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZXhwIjoxNzA5OTk5OTk5fQ.signature"

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title (required)
          example: "Implement user authentication"
        description:
          type: string
          maxLength: 5000
          nullable: true
          description: Optional detailed task description
          example: "Add JWT-based auth to the API"

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Updated task title
          example: "Implement user authentication (updated)"
        description:
          type: string
          maxLength: 5000
          nullable: true
          description: Updated task description
          example: "Add JWT-based auth with refresh tokens"
        completed:
          type: boolean
          description: Updated completion status
          example: true

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 1
        user_id:
          type: string
          format: uuid
          description: Owner user ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: Task title
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          description: Task description (optional)
          example: "Write comprehensive API docs"
        completed:
          type: boolean
          description: Completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp (ISO 8601)
          example: "2026-02-08T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          example: "2026-02-08T10:30:00Z"

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Validation error"
        field_errors:
          type: object
          additionalProperties:
            type: string
          description: Field-specific validation errors
          example:
            title: "Title cannot be empty"

  responses:
    Unauthorized:
      description: Missing, invalid, or expired JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Missing or invalid token"

    Forbidden:
      description: Valid token but user_id mismatch (cross-user access attempt)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Access denied: user_id mismatch"

    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Internal server error"